+++ src/Fauna/herbivore_base.cpp
@@ -23,7 +23,8 @@ HerbivoreBase::HerbivoreBase(const int age_days, const double body_condition,
       forage_gross_energy(forage_gross_energy),
       energy_budget(body_condition * get_max_fatmass(),  // initial fat mass
                     get_max_fatmass(),                   // maximum fat mass
-                    hft->digestion_anabolism_coefficient,
+                    hft->body_fat_gross_energy * hft->digestion_k_maintenance /
+                        hft->digestion_k_fat,
                     hft->body_fat_gross_energy),
       get_forage_demands_per_ind(hft, sex),
       today(-1),  // not initialized yet; call simulate_day() first
@@ -61,7 +62,9 @@ HerbivoreBase::HerbivoreBase(std::shared_ptr<const Hft> hft, const Sex sex,
       forage_gross_energy(forage_gross_energy),
       breeding_season(hft->breeding_season_start, hft->breeding_season_length),
       energy_budget(get_hft().body_fat_birth * get_hft().body_mass_birth,
-                    get_max_fatmass(), hft->digestion_anabolism_coefficient,
+                    get_max_fatmass(),
+                    hft->body_fat_gross_energy * hft->digestion_k_maintenance /
+                        hft->digestion_k_fat,
                     hft->body_fat_gross_energy),
       get_forage_demands_per_ind(hft, sex),
       body_condition_gestation(get_hft().reproduction_gestation_length * 30) {}